Roles

+++ Modules/UpendoUserManaged/ViewModels/Pagination.cs
@@ -7,12 +7,12 @@ namespace Upendo.Modules.UpendoUserManaged.ViewModels
 {
     public struct DataTableResponse<T>
     {
-        public int Take { get; set; }
+        public double Take { get; set; }
         public int Skip { get; set; }
         public int Page { get; set; }
         public int GoToPage { get; set; }
-        public int PagesTotal { get; set; }
-        public int RecordsTotal { get; set; }
+        public double PagesTotal { get; set; }
+        public double RecordsTotal { get; set; }
         public int RecordsFiltered { get; set; }
         public string Search;
         public string OrderBy;
